pyproject.toml
[tool.poetry]
name = "-devcontainer"
version = "0.1.0"
description = ""
authors = ["Mathew Werber <matwerber@gmail.com>"]
readme = "README.md"
packages = [{include = "devcontainer"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.33.5"
boto3-stubs = {extras = ["all"], version = "^1.33.5"}
botocore-stubs = {extras = ["all"], version = "^1.33.5"}
boto-session-manager = "^1.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
